Added completed percentage in SubjectsController

+++ Models/Serializer/SubjectsSerializer.cs
@@ -5,7 +5,7 @@ public class Subject
     public int Id { get; set; }
     public string Name { get; set; }
     public int ClassId{get; set;}
-    public int? Percentage { get; set; }
+    public double? Percentage { get; set; }
     public string Logo { get; set; }
 }
